requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2.5
django-crontab==0.7.1
django-filter==23.2
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2-binary==2.9.7
pytz==2023.3.post1
redis==5.0.0
sqlparse==0.4.4
